requirements.txt
# Deps that are Colab/local-safe (Colab already ships torch+torchvision with CUDA).
# For local install, install torch + torchvision FIRST via `requirements-torch-cuda.txt`.

timm>=1.0.0
torchmetrics>=1.4.0
torchinfo>=1.8.0

# Data / augmentation
albumentations>=1.4.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0

# Science stack
numpy>=1.26
scipy>=1.12
scikit-learn>=1.4
pandas>=2.2

# Plotting
matplotlib>=3.8
seaborn>=0.13

# UX
tqdm>=4.66
ipywidgets>=8.1
jupyterlab>=4.1

# Interpretability
grad-cam>=1.5.0

# Deployment / optimization
onnx>=1.16
onnxruntime>=1.17
